Use string columns for JSON data due to polymorphic deserialization issues
System.Text.Json expected $type to be the first property in a JSON object but PostgreSQL jsonb column does not return the original string, rather some new rendering of it. All the properties are included but just not in the correct order. This can be solved later on if there is real need for jsonb columns.